feature #3782 extend to return disk and snapshot info for 'dev' type via external program
pass the dev parameter to external script available in .../vmm/kvm:
> ../../vmm/kvm/poll_disk_info.#{systemds_id} #{dev} #{type}

The script returns XML with the used size in this device and if there are snapshots the size per snapshot id.
The XML is parsed and the desired structure is filled up.

Here is the XML schema of the returned xml:

        <?xml version="1.0" encoding="UTF-8" ?>
        <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
          <xs:element name="disk">
            <xs:element name="size" type="xs:positiveInteger"/>
            <xs:element name="snapshot" minOccurs="0" maxOccurs="unbounded">
              <xs:complexType>
                <xs:sequence>
                  <xs:element name="id" type="xs:positiveInteger"/>
                  <xs:element name="size" type="xs:positiveInteger"/>
                </xs:sequence>
              </xs:complexType>
            </xs:element>
          </xs:element>
        </xs:schema>

The #{systemds_id} was got from VM metadata <system_datastore>or if the device is block device or file - from its path

I think it is better to add for disk_id its TM_MAD in the VM metadata field and then we can call poll_disk_info for the given TM_MAD